core/requirements.txt
ipykernel
pytesseract
pillow
pymupdf
pymysql
python-dotenv
chromadb
sentence-transformers
ollama
langchain-core
langchain-text-splitters
groq
langsmith